+++ src/graphql/block_network.rs
@@ -119,7 +119,8 @@ impl BlockNetworkMutation {
         .or(RoleGuard::new(Role::SecurityAdministrator))")]
     async fn apply_block_networks(&self, ctx: &Context<'_>) -> Result<Vec<String>> {
         let db = ctx.data::<Arc<Store>>()?;
-        let serialized_networks = bincode::serialize(&get_block_networks(db)?)?;
+        let serialized_networks =
+            bincode::DefaultOptions::new().serialize(&get_block_networks(db)?)?;
         let agent_manager = ctx.data::<BoxedAgentManager>()?;
         agent_manager
             .broadcast_block_networks(&serialized_networks)