+++ CHANGELOG.md
@@ -5,6 +5,14 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Change to serialize values with `bincode::DefaultOptions::new().serialize()`
+  instead of `bincode::serialize()` when broadcasting
+  `internal networks, allow/block list`
+
 ## [0.10.0] - 2023-05-31
 
 ### Added